correct api error handling

+++ backend/src/menus/menus.service.ts
@@ -1,4 +1,4 @@
-import { Injectable } from '@nestjs/common';
+import { BadRequestException, Injectable, NotFoundException, UnauthorizedException } from '@nestjs/common';
 import { MenuItem } from '@prisma/client';
 
 import { PrismaService } from "../prisma/prisma.service";
@@ -12,9 +12,15 @@ export class MenusService {
 	}
 
 	async getMenuItem(id: number): Promise<MenuItem> {
-		return await this.prisma.menuItem.findUnique({
+		const menuItem = await this.prisma.menuItem.findUnique({
 			where: { id }
 		});
+
+		if (!menuItem) {
+			throw new NotFoundException('Menu item not found');
+		}
+
+		return menuItem;
 	}
 
 	async addMenuItem(name: string, description: string, price: number): Promise<MenuItem> {
@@ -28,7 +34,9 @@ export class MenusService {
 	}
 
 	async updateMenuItem(id: number, name?: string, description?: string, price?: number): Promise<MenuItem> {
-		if (!id) return null;
+		if (!id) {
+			throw new BadRequestException('No ID provided');
+		}
 		
 		let updateData = {};
 
@@ -36,13 +44,34 @@ export class MenusService {
 		if (description) updateData['description'] = description;
 		if (price !== undefined && price !== null) updateData['price'] = +price;
 
+		// check if menu item exists
+		const menuItem = await this.prisma.menuItem.findUnique({
+			where: { id }
+		});
+
+		if (!menuItem) {
+			throw new NotFoundException('Menu item not found');
+		}
+
 		return await this.prisma.menuItem.update({
 			where: { id },
 			data: updateData
 		});
 	}
 
-	deleteMenuItem(id: number) {
+	async deleteMenuItem(id: number) {
+		if (!id) {
+			throw new BadRequestException('No ID provided');
+		}
+
+		const menuItem = await this.prisma.menuItem.findUnique({
+			where: { id }
+		});
+
+		if (!menuItem) {
+			throw new NotFoundException('Menu item not found');
+		}
+
 		return this.prisma.menuItem.delete({
 			where: { id }
 		});
